+++ packages/common/src/index.js
@@ -62,7 +62,7 @@ export { simplePubsub } from './Utils/simplePubsub';
 export { getModalStyles } from './Utils/getModalStyles';
 export { mergeStyles } from './Utils/mergeStyles';
 export { default as normalizeInitialState } from './Utils/normalizeInitialState';
-export { validate } from './Utils/component-data-validator';
+export { validate } from './Utils/data-schema-validator';
 export { isHexColor } from './Components/ColorPicker/utils';
 
 export {