+++ packages/plugin-text-color/src/constants.js
@@ -5,10 +5,6 @@ export const DEFAULT_COLOR = '#000000';
 export const PANEL_WIDTH = 216;
 export const PANEL_HEIGHT = 116;
 
-export const DEFAULT_COLOR_TO_STYLE = color => color;
-
-export const DEFAULT_STYLE_TO_COLOR = style => style;
-
 export const DEFAULT_STYLE_SELECTION_PREDICATE = style => isHexColor(style);
 
 export const DEFAULT_STYLE_FN = style => (isHexColor(style) ? { color: style } : {});