+++ packages/plugin-text-color/src/toolbar/TextColorPanel.jsx
@@ -1,26 +1,31 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import isFunction from 'lodash/isFunction';
 import { Modifier, EditorState } from '@wix/draft-js';
 import { ColorPicker, getSelectionStyles } from 'wix-rich-content-common';
+import { DEFAULT_COLOR, DEFAULT_STYLE_SELECTION_PREDICATE } from '../constants';
+
 import {
-  DEFAULT_PALETTE,
-  DEFAULT_COLOR,
-  DEFAULT_COLOR_TO_STYLE,
-  DEFAULT_STYLE_TO_COLOR,
-  DEFAULT_STYLE_SELECTION_PREDICATE,
-} from '../constants';
+  extractColor,
+  extractPalette,
+  extractSchemeAttributes,
+  validateColorScheme,
+} from '../color-scheme-utils';
 
 export default class TextColorPanel extends Component {
   constructor(props) {
     super(props);
     const styleSelectionPredicate =
       props.settings.styleSelectionPredicate || DEFAULT_STYLE_SELECTION_PREDICATE;
-    const styleToColor = props.settings.styleToColor || DEFAULT_STYLE_TO_COLOR;
+    if (props.settings.colorScheme && !validateColorScheme(props.settings.colorScheme)) {
+      console.error('Error: colorScheme is not valid'); // eslint-disable-line no-console
+    }
     const currentColors = getSelectionStyles(styleSelectionPredicate, props.editorState);
     this.state = {
       currentColor:
-        currentColors.length > 0 ? styleToColor(currentColors[0]).toUpperCase() : DEFAULT_COLOR,
+        currentColors.length > 0
+          ? extractColor(props.settings.colorScheme, currentColors[0])
+          : DEFAULT_COLOR,
+      currentSchemeColor: currentColors[0],
       userColors: props.settings.getUserColors() || [],
     };
     this.setColor = this.setColor.bind(this);
@@ -32,12 +37,11 @@ export default class TextColorPanel extends Component {
   }
 
   setColor(color) {
-    const colorToStyle = this.props.settings.colorToStyle || DEFAULT_COLOR_TO_STYLE;
-    const style = colorToStyle(color);
     if (color !== this.state.currentColor) {
-      this.applyInlineColorStyle(style);
+      this.applyInlineColorStyle(color);
       this.setState({
-        currentColor: color,
+        currentColor: extractColor(this.props.settings.colorScheme, color),
+        currentSchemeColor: color,
       });
     }
     this.props.closeModal();
@@ -70,12 +74,14 @@ export default class TextColorPanel extends Component {
 
   render() {
     const { theme, settings, t, setKeepToolbarOpen, isMobile } = this.props;
-    const palette = isFunction(settings.getPaletteColors)
-      ? settings.getPaletteColors()
-      : DEFAULT_PALETTE;
+    const { colorScheme } = settings;
+    const palette = extractPalette(colorScheme);
+    const schemeAttributes = extractSchemeAttributes(colorScheme);
     const { onCustomPickerToggle, onCustomColorPicked } = settings;
     return (
       <ColorPicker
+        schemeAttributes={schemeAttributes}
+        schemeColor={this.state.currentSchemeColor}
         color={this.state.currentColor}
         palette={palette.slice(0, 6)}
         userColors={this.state.userColors.slice(0, 17)}
@@ -100,13 +106,11 @@ TextColorPanel.propTypes = {
   uiSettings: PropTypes.object,
   settings: PropTypes.shape({
     onColorAdded: PropTypes.func.isRequired,
-    getPaletteColors: PropTypes.func,
+    colorScheme: PropTypes.object,
     getUserColors: PropTypes.func,
     onCustomPickerToggle: PropTypes.func,
     onCustomColorPicked: PropTypes.func,
     styleSelectionPredicate: PropTypes.func,
-    colorToStyle: PropTypes.func,
-    styleToColor: PropTypes.func,
   }).isRequired,
   setKeepToolbarOpen: PropTypes.func,
   closeModal: PropTypes.func.isRequired,