+++ CHANGELOG.md
@@ -13,10 +13,13 @@
   - [#282](https://github.com/wix-incubator/rich-content/pull/282) new plugin added
 - `common`
   - [#294](https://github.com/wix-incubator/rich-content/pull/294) ColorPicker now accepts `onCustomPickerToggle`, `onCustomColorPicked` props for higher customability
+  - [#301](https://github.com/wix-incubator/rich-content/pull/301) ColorPicker now accepts `schemeColor`, `schemeAttributes` props for scheme color support
 - `text-color`
   - [#294](https://github.com/wix-incubator/rich-content/pull/294) plugin config now accepts `onCustomPickerToggle`, `onCustomColorPicked` handlers
+  - [#301](https://github.com/wix-incubator/rich-content/pull/301) plugin config now accepts `colorScheme` mapping instead of `getPaletteColors` and converter functions
 ### :bug: Bug Fix
 - `text-color`
+  - [#301](https://github.com/wix-incubator/rich-content/pull/301) viewer decorator does not override other inline styles
   - [#300](https://github.com/wix-incubator/rich-content/pull/300) missing `getPaletteColors` won't crash
   - [#300](https://github.com/wix-incubator/rich-content/pull/300) pop-up is opened as internal modal rather external one
   - [#300](https://github.com/wix-incubator/rich-content/pull/300) pop-up is themable